% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_plotters.R
\name{agerange_map}
\alias{agerange_map}
\title{#' Plot AgeRange variable}
\usage{
agerange_map(df)
}
\arguments{
\item{df}{Dataframe with the results from searching clinicaltrials.gov}
}
\description{
This function takes the set dataframe and plots the result of collapsing MinAge 
and MaxAge variables as a treemap.
}
\keyword{AgeRange}
\keyword{MaxAge}
\keyword{MinAge}
\keyword{plot}
