% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armify.R
\name{armify}
\alias{armify}
\title{To Intervention Arms Table}
\usage{
armify(df)
}
\arguments{
\item{df}{a dataframe to manipulate}
}
\description{
Takes data frame with studies restults and returnsdataframe with one row per arm
 with unioque study identifier.
armify()
}
\examples{
armify(df)
}
