% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_app_input.R
\name{check_app_input}
\alias{check_app_input}
\title{Parse Check App Input}
\usage{
check_app_input(data)
}
\arguments{
\item{data}{list of 2 or dataframe}
}
\description{
This function parses the app input to create a search key
}
\examples{
check_app_input(data)
}
\keyword{check}
\keyword{data}
\keyword{input}
