% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_app_input.R
\name{check_df}
\alias{check_df}
\title{Check DF}
\usage{
check_df(data)
}
\arguments{
\item{data}{list of 2 or dataframe}
}
\description{
This function checks the dataframe input
}
\examples{
check_df(data)
}
\keyword{check}
\keyword{data}
\keyword{input}
