% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_app_input.R
\name{check_key}
\alias{check_key}
\title{Check Key}
\usage{
check_key(data)
}
\arguments{
\item{data}{list of 2 or dataframe}
}
\description{
This function checks the key input
}
\examples{
check_key(data)
}
\keyword{check}
\keyword{data}
\keyword{input}
