% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_app_input.R
\name{clean_app_input}
\alias{clean_app_input}
\title{Clean Input for Data Explorer}
\usage{
clean_app_input(df)
}
\arguments{
\item{df}{a dataframe to manipulate}
}
\description{
clean-app_input()
}
\examples{
clean_app_input(search_expr = 'psoriatic arthritis',fields = for_explorer, max_studies = 300)
}
