% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_plotters.R
\name{condition_map}
\alias{condition_map}
\title{Plot Condition variable}
\usage{
condition_map(df)
}
\arguments{
\item{df}{Dataframe with the results from searching clinicaltrials.gov}
}
\description{
This function takes the set dataframe and plots the condition variable as a treemap.
}
\keyword{Condition}
\keyword{plot}
