% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_plotters.R
\name{gender_map}
\alias{gender_map}
\title{Plot Gender variable}
\usage{
gender_map(df)
}
\arguments{
\item{df}{Dataframe with the results from searching clinicaltrials.gov}
}
\description{
This function takes the set dataframe and plots the gender variable as a treemap.
}
\keyword{gender}
\keyword{plot}
