% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_explorer}
\alias{launch_explorer}
\title{Launch Data Explorer Shiny App}
\usage{
launch_explorer(df = df)
}
\arguments{
\item{df}{Dataframe obtained through set_app_input()}
}
\description{
This function takes a dataframe as parameter and launches the explorer application.
}
\examples{
df <- set_app_input("psoriatic arthritis")
launch_explorer(df)
}
\keyword{app}
\keyword{launch}
