% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_MESH_plotters.R
\name{meshterm_dmc_barplot}
\alias{meshterm_dmc_barplot}
\title{Plot MESH Term by}
\usage{
meshterm_dmc_barplot(df)
}
\arguments{
\item{df}{Dataframe with the results from searching clinicaltrials.gov}
}
\description{
Plot MESH Term by
}
\keyword{plot}
