% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_cols_to_factors.R
\name{named_cols_to_factors}
\alias{named_cols_to_factors}
\title{named_cols_to_factors}
\usage{
named_cols_to_factors(df, fct_cols = fct_cols)
}
\arguments{
\item{df}{a dataframe}

\item{fct_cols}{a character vector, with the names of columns in the dataframe}
}
\description{
Carries out class transform to factor.
Enhances datatable usage by providing ready levels on click of the filter boxes.
}
\details{
named_cols_to_factors()
}
