% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_plotters.R
\name{overallstatus_map}
\alias{overallstatus_map}
\title{#' Plot OverallStatus variable}
\usage{
overallstatus_map(df)
}
\arguments{
\item{df}{Dataframe with the results from searching clinicaltrials.gov}
}
\description{
This function takes the set dataframe and plots the overall status variable as a treemap.
}
\keyword{OverallStatus}
\keyword{plot}
