% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_plotters.R
\name{phase_map}
\alias{phase_map}
\title{#' Plot Phase variable}
\usage{
phase_map(df)
}
\arguments{
\item{df}{Dataframe with the results from searching clinicaltrials.gov}
}
\description{
This function takes the set dataframe and plots the phase variable as a treemap.
}
\keyword{Phase}
\keyword{plot}
