% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap_plotters.R
\name{studytype_map}
\alias{studytype_map}
\title{Plot Study Type variable}
\usage{
studytype_map(df)
}
\arguments{
\item{df}{Dataframe with the results from searching clinicaltrials.gov}
}
\description{
This function takes the set dataframe and plots the study type variable as a treemap.
}
\keyword{StudyType}
\keyword{plot}
