% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_arms.R
\name{to_arms}
\alias{to_arms}
\title{To Intervention Arms Table}
\usage{
to_arms(df)
}
\arguments{
\item{df}{a dataframe to manipulate}
}
\description{
Takes data frame with studies restults and returns dataframe with one row per arm
}
\details{
to_arms()
}
\examples{
to_arms(df)
}
