% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_nodes.R
\name{to_nodes}
\alias{to_nodes}
\title{To Nodes}
\usage{
to_nodes(edges)
}
\arguments{
\item{df}{a dataframe to manipulate}
}
\description{
Takes data frame with edges and extracts a node dataframe
to_nodes(edges)
}
\examples{
to_nodes(edges)
}
