% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_wide_arms.R
\name{to_wide_arms}
\alias{to_wide_arms}
\title{To Interventions Table}
\usage{
to_wide_arms(df)
}
\arguments{
\item{df}{a dataframe to manipulate}
}
\description{
Takes data frame with studies restults and returns dataframe with one row per study,
and one intervention per column
to_wide_arms()
}
\examples{
to_wide_arms(df)
}
