% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_explorer}
\alias{launch_explorer}
\title{Launch Data Explorer Shiny App}
\usage{
launch_explorer(df)
}
\arguments{
\item{df}{Dataframe obtained through set_app_input()}
}
\description{
This function takes a dataframe as parameter and launches the explorer application.
}
\examples{
launch_explorer(set_app_input(search_expr = 'psoriatic arthritis',fields = for_explorer, max_studies = 300))
}
\keyword{app}
\keyword{launch}
