% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launch_explorer}
\alias{launch_explorer}
\title{Launch Data Explorer Shiny App}
\usage{
launch_explorer(df)
}
\arguments{
\item{df}{Dataframe obtained through set_app_input()}
}
\description{
This function takes one parameter
}
\examples{
launch_explorer()
}
\keyword{app}
\keyword{launch}
