% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_app_input.R
\name{set_app_input}
\alias{set_app_input}
\title{Set Input for Data Explorer}
\usage{
set_app_input(search_expr, fields = for_explorer, max_studies = 500)
}
\arguments{
\item{search_expr}{A string following the querying guidelines at
https://clinicaltrials.gov/api/gui/ref/syntax and https://clinicaltrials.gov/api/gui/ref/expr}

\item{fields}{A character vector with field names. Default list for_explorer is optimised for
data explorer app performance.}

\item{max_studies}{A number indicating how many studies out of all that
match the search expression will be returned.}
}
\description{
This function takes a search expression and character vector with field names
and returns one dataframe. This dataframe which must be named "df" is used by the data explorer app called by
launch_explorer()
}
\examples{
set_app_input(expr = 'psoriatic arthritis',fields = for_explorer, max_studies = 300)
}
\keyword{app}
\keyword{input}
\keyword{set}
